requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pandas>=2.2,<3.0
gspread>=6.0,<7.0
google-auth>=2.29,<3.0
python-multipart>=0.0.9
